git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/branches/geant4/_symbols/geant4-09-05_branch@54915 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/externals/expat/src/xmlparse.cc
@@ -1890,7 +1890,7 @@ XML_ErrorString(enum XML_Error code)
     XML_L("reserved prefix (xmlns) must not be declared or undeclared"),
     XML_L("prefix must not be bound to one of the reserved namespace names")
   };
-  if (code > 0 && code < sizeof(message)/sizeof(message[0]))
+  if (code > 0 && unsigned(code) < sizeof(message)/sizeof(message[0]))
     return message[code];
   return NULL;
 }
